refactor(reviewer-bot): extract lifecycle handlers
